refactor: extract shared SectionNavDropdown for settings/fitness sub-nav
Address review-round findings:

- Extract the duplicated dropdown sub-nav (trigger + menu + active-tab
  resolution) from the settings and fitness layouts into a single shared
  lib/components/section-nav-dropdown.tsx, so future tweaks happen once
  instead of being copy-pasted across both layouts.
- Harden the layout tests: scope the opened-menu assertions to the
  `menu` role via within(menu) so they fail loudly if items ever render
  outside an open dropdown.
- Add a fitness nested-path test (/fitness/files/abc123 -> Files) mirroring
  the settings nested-path coverage, protecting the most-specific active-tab
  resolution on the fitness side too.
