+++ app/(timeline)/settings/layout.test.tsx
@@ -66,16 +66,17 @@ describe('Settings Layout', () => {
 
   // Open the Radix menu the same way the rest of the suite does (keyboard, since
   // jsdom has no pointer layout), then assert the items the rail used to expose.
+  // Scope to the open menu so the test fails loudly if items ever render outside
+  // an opened dropdown.
   it('renders every section as a menu item when the dropdown is opened', async () => {
     ;(usePathname as jest.Mock).mockReturnValue('/settings')
     renderLayout()
 
     fireEvent.keyDown(screen.getByRole('button'), { key: 'ArrowDown' })
 
-    expect(
-      await screen.findByRole('menuitem', { name: 'General' })
-    ).toBeInTheDocument()
+    const menu = await screen.findByRole('menu')
     for (const label of [
+      'General',
       'Account',
       'Media',
       'Notifications',
@@ -83,7 +84,9 @@ describe('Settings Layout', () => {
       'Muted accounts',
       'Sessions'
     ]) {
-      expect(screen.getByRole('menuitem', { name: label })).toBeInTheDocument()
+      expect(
+        within(menu).getByRole('menuitem', { name: label })
+      ).toBeInTheDocument()
     }
   })
 
@@ -93,11 +96,12 @@ describe('Settings Layout', () => {
 
     fireEvent.keyDown(screen.getByRole('button'), { key: 'ArrowDown' })
 
+    const menu = await screen.findByRole('menu')
     expect(
-      await screen.findByRole('menuitem', { name: 'Account' })
+      within(menu).getByRole('menuitem', { name: 'Account' })
     ).toHaveAttribute('aria-current', 'page')
     expect(
-      screen.getByRole('menuitem', { name: 'General' })
+      within(menu).getByRole('menuitem', { name: 'General' })
     ).not.toHaveAttribute('aria-current')
   })
 })
